requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pyyaml>=6.0.1
transformers>=4.43.0
torch>=2.2.0
shap>=0.43.0
numpy>=1.24.0
requests>=2.31.0
streamlit>=1.35.0
